BLD: Refactor io into folders
