--- client/io/IOBase.cc
@@ -1,12 +1,0 @@
-/*
-** This file is part of eOn.
-**
-** SPDX-License-Identifier: BSD-3-Clause
-**
-** Copyright (c) 2010--present, eOn Development Team
-** All rights reserved.
-**
-** Repo:
-** https://github.com/TheochemUI/eOn
-*/
-#include "client/io/IOBase.hpp"
