+++ client/io/writers/con/ConBase.hpp
