+++ client/io/writers/con/meson.build
@@ -1,0 +1,13 @@
+_linkto += library(
+    'ecio_con',
+    sources: [
+      'ConBase.cc',
+      'ConWriter.cc',
+      'ConvelWriter.cc',
+    ],
+    include_directories: _incdirs,
+    dependencies: _deps,
+    link_with: _linkto,
+    cpp_args: _args,
+    install: true,
+)
