+++ client/io/writers/meson.build
@@ -1,0 +1,3 @@
+subdir('con')
+subdir('tibble')
+subdir('xyz')
