+++ client/io/writers/tibble/meson.build
@@ -1,0 +1,9 @@
+_linkto += library(
+    'ecio_tibble',
+    'TibbleWriter.cc',
+    include_directories: _incdirs,
+    dependencies: _deps,
+    link_with: _linkto,
+    cpp_args: _args,
+    install: true,
+)
