+++ client/io/writers/xyz/meson.build
@@ -1,0 +1,9 @@
+_linkto += library(
+    'ecio_xyz',
+    'XYZWriter.cc',
+    include_directories: _incdirs,
+    dependencies: _deps,
+    link_with: _linkto,
+    cpp_args: _args,
+    install: true,
+)
